+++ changelog.md
@@ -1,4 +1,4 @@
-# Upcoming
+# 1.3.1.0-beta
 * Added advancements for obtaining your first rune, removing a rune with a grindstone and learning a curse
 * Veining and Tunneling are now applicable to shears